Step3 빼빼로 그리기 & 순서도이미지파일로 다시 올려놓음
