+++ CodeStarterCamp_Week1/Icecream.swift
@@ -1,0 +1,20 @@
+//
+//  Icecream.swift
+//  CodeStarterCamp_Week1
+//
+//  Created by Channoori Park on 2022/10/25.
+//
+
+import Foundation
+
+func drawIcecreamCream() {
+    for _ in 1...8 {
+        print("***********")
+    }
+}
+
+func drawIcecreamStick() {
+    for _ in 1...4 {
+        print("    | |    ")
+    }
+}
