+++ CodeStarterCamp_Week1/main.swift
@@ -8,18 +8,15 @@
 
 import Foundation
 
-func drawIcecreamCream() {
-    for _ in 1...8 {
-        print("***********")
-    }
-}
-
-func drawIcecreamStick() {
-    for _ in 1...4 {
-        print("    | |    ")
-    }
-}
+//drawIcecreamCream()
+//drawIcecreamStick()
 
 
-drawIcecreamCream()
-drawIcecreamStick()
+makePepero(10, stickHeight: 4, body: "***")
+print("\n\n")
+makePepero(12, stickHeight: 4, body: "***", toping: "&")
+print("\n\n")
+makePepero(12, stickHeight: 6, body: "***", toping: "#")
+print("\n\n")
+makePepero(6, stickHeight: 4, body: "|0|")
+print("\n\n")
