--- flowChart/아이스크림.drawio
@@ -1,1 +1,0 @@
-<mxfile host="app.diagrams.net" modified="2022-10-21T16:59:52.354Z" agent="5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/106.0.0.0 Safari/537.36" etag="rWhHBvb_mDyeibvSsb04" version="20.5.0" type="google"><diagram id="C5RBs43oDa-KdzZeNtuy" name="Page-1">7Vtfc5s4EP80vNyMM4AMhkdwnPSuuUxv0rtcnm5kUGwagSjg2O6nrwSS+W+TONikl8w0kcRqWWn3t7taUQlM/c11BMPln8RFWFLlReS5EriUVFWh/+hACBeoNMAo7rwfYlDmoyvPRXGJMCEEJ15YHnRIECAnKY3BKCLrMtkjwXUx7hyIUW303nOTZTZqqJN8/BPyFkvxIkU3syc+FMRc8HgJXbIuDIGZBKYRIUnW8jdThNnOiH25/317j2+e9Os//oq/w7/tz19v/xllzK5eMmW3hAgFyduyVjPWzxCv+H7xtSZbsYERWQUuYkxkCdjLxMe0qdDmN5QkW65fuEoIHSJRsiQLEkB8Q0jI6R5JkHAyhfVR4FpMj7Q/x8R5yoauPIz5O2iP0xu0FycRedrpjjHYKYIRYzhH2IbO0yIVdEowieijgASIsXKpMfC15MLN8lEbbbzkX8brQuO9B/4a1r7cCJlYZ7sTMIm2hUms+1B8lk9Le2JetrdMpoptHlAsp4vJKnL41MW1cf/o336xvn39z/6M9C8jeDMCGscTjBYo2Ue4M18KakR8RKWk8yKEYeI9l6WDHG+LHd1u6hfiUblVeZNR7JCy5QxEX7DIFsBnFU31EKNJhVG2wBoj2iisJx9KkdCMiubt0Q/D4rBllYHTCI0GCBUAxoBzBX0Ps234hPAzSjwHVhCl8r4wfEkFcvpzAnuTu5qbdqS9lRT8Um3KDcrUMZXXntPGgjWk2VQybMmcsoY5k0xFkNA35lQ1G8CYRi2m6/XSS9BdCNPtWtMwWdYklwBFCdq8QhX1nRNc9DJQRgrgA+s81u1guCzEOUPuabeNcyDll3LiHUE1OSemQB1Ts0vJ1CTLSLFkSCY1BNmN4HoaIeinuLqUjDEd5GCzxukUXTJMPtdM55oTyZoyMlX/vmK5lf1b/pMPcka2nqKWspMlSxOM0pH0HRcH8pmBAFfROuJW7wu3Sl2lDzRNPweYBwrMSUdgKi2qF6rmGnxtsiVIyONjjHpJf5RhpD/v3Q7AsR66OTkey2XXAaqhvOfkWDFr1nFLGu3jhh3NyjqF2FsEtO1QVSCartrMtdKkFlv8ge+5bmY+KPZ+wHnKj2k1ZOtJV6jZknbZkgW3psvt6TGvM/BXSbuzxwGnvwc5rV5fvlAMwWtb4jxcZzCpadspBHSbBXSDRVsbsI4Iu2l0pvFekWzasNhvaqbgqh6Nl8Sfr+LDkbjsTnqLy5pWUo/alE+LY3YxLityb4F5fA7n27cTbQuSfZcgWpyjFUVwWyDj3qb1Pcq4/B7hlNsqGlV6IVfnCkiZnjYyid/Wsyt7sD6RTCqCrJQzeQp3mnGP63l3zWaHAe9K2g0mZ0+7jUHkWoMvNSldS5vKWc/FQsyGYlNYUqo4xvp0PR5NfCz6VA432cGWbfLoGUYepH+DlY8iz8lIAhL5EDdQIRgnIxjTdgthnEQocZb1p9gL0EjYevUhxR9lm+VmVroAGGWn78I53MEwjkurCpUigaixtRQI7qipPR1VIBBbHId0+U27fEBcKwwxGtFEMPU47ihO3VR9BaoOfea3gnkc8hJE45CoHGby8MmTVGb2+x0I+4oyy45R2FovHUQMANXQbTaleKeMAepHDOgWA7qevNWzxgC1fkBuq6T9/07Ima3vOSHLpjjwDPdMDBrK3x+APQKwQD0SsK86xGmVSKDr+w9lVXpTO8GhDIxrtvZRbivhsNWZjOQL1RR3+cP1Jkq9xBYXUuK0xDb+dUpsAFRQpzbkX+OG/Ku3K2ug1PbslHfWUn5jnd9fN99ZD97jH515dbvzGHf8IOilEQFUym7GgTJdlV7I1WtEUPf5ixeW6erX48PwEZVIO2k6o520TgfUesrHNlFhu1zzHl2/CzrPpYbZ5HFPu5v1MllWaU6LTHV3PJSvNM6+b+/m66rBRyrQ+XOOs368CNqcfQtSXvR10/5vp99Ihf1HB9BjdKDd/Av/LILn/wkCzH4C</diagram></mxfile>
